Loyalty Phase 3d: run promotion engine natively (drop proxy hop)
Move the promotion discount engine out of apps/loyalty into
@celsius/shared so POS (backoffice) and pickup/order checkout evaluate
the discount stack in-process against the shared Supabase, instead of
proxying to loyalty.celsiuscoffee.com/api/promotions/{evaluate,apply}.

- packages/shared/src/loyalty/promo-engine.ts: lift evaluateCart +
  recordApplications (and helpers) verbatim, swapping the hard-wired
  loyalty supabaseAdmin for an injected SupabaseClient. Discount maths
  is unchanged, including the 60s promo/tier caches. Imported via
  subpath (not the barrel) since it pulls in node `crypto`, mirroring
  the ./otp precedent so it never lands in a client bundle.
- apps/loyalty promotions.ts: thin shim binding the loyalty app's own
  supabaseAdmin to the shared engine; routes unchanged (old builds safe).
- apps/order promotions.ts + apps/backoffice evaluate-promotions: native
  first, with auto-fallback to the loyalty proxy on error and a
  LOYALTY_PROMO_USE_PROXY=1 kill-switch to revert without a deploy. Tier
  and first-order post-steps unchanged.
